requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain-community>=0.0.10
langchain-core>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pypdf2>=3.0.0
python-multipart>=0.0.6
